requirements.txt
onvif-zeep
zeep
